FS-28: Made sure the temporary files get removed when an upload or update fails.

+++ src/main/java/org/fides/server/ClientFileConnector.java
@@ -26,7 +26,7 @@ import com.google.gson.JsonObject;
 
 /**
  * A class for handling the sending and receiving of files
- *
+ * 
  * @author Thijs
  */
 public class ClientFileConnector {
@@ -40,7 +40,7 @@ public class ClientFileConnector {
 
 	/**
 	 * Constructor for ClientFileConnector
-	 *
+	 * 
 	 * @param userFile
 	 *            The logged in user's userFile
 	 */
@@ -50,7 +50,7 @@ public class ClientFileConnector {
 
 	/**
 	 * Downloads the keyfile of the currently logged in user.
-	 *
+	 * 
 	 * @param outputStream
 	 *            The stream which contains the keyfile.
 	 * @return Wether the writing of the keyfile to the stream was successful
@@ -72,7 +72,7 @@ public class ClientFileConnector {
 
 	/**
 	 * Downloads a file by writing it to the outputstream
-	 *
+	 * 
 	 * @param fileRequest
 	 *            The Json request which contains the file's location
 	 * @param outputStream
@@ -104,7 +104,7 @@ public class ClientFileConnector {
 
 	/**
 	 * Copies an errormessage to the outputstream
-	 *
+	 * 
 	 * @param errorMessage
 	 *            The message to copy
 	 * @param outputStream
@@ -124,7 +124,7 @@ public class ClientFileConnector {
 	/**
 	 * Uploads a new file with the inputStream as its content. Generates a new file on the server to fill with the
 	 * stream and returns the file's server location to the client through the outputstream.
-	 *
+	 * 
 	 * @param inputStream
 	 *            The content of the file
 	 * @param outputStream
@@ -134,9 +134,11 @@ public class ClientFileConnector {
 	public boolean uploadFile(DataInputStream inputStream, DataOutputStream outputStream) {
 		String location = FileManager.createFile();
 		File file = new File(PropertiesManager.getInstance().getDataDir(), location);
+		boolean uploadSuccessful = false;
 		// Check if the file was created correctly (should always be true)
 		if (file.exists()) {
-			try {
+			try (InputStream virtualInputStream = new VirtualInputStream(inputStream);
+				OutputStream fileOutputStream = new FileOutputStream(file)) {
 				// Return the location on the server where the new file will be written
 				JsonObject returnJobj = new JsonObject();
 				returnJobj.addProperty(Responses.SUCCESSFUL, true);
@@ -144,8 +146,6 @@ public class ClientFileConnector {
 				outputStream.writeUTF(new Gson().toJson(returnJobj));
 
 				// Put the inputstream received from the user into a temporary file
-				InputStream virtualInputStream = new VirtualInputStream(inputStream);
-				OutputStream fileOutputStream = new FileOutputStream(file);
 				IOUtils.copy(virtualInputStream, fileOutputStream);
 				fileOutputStream.flush();
 				fileOutputStream.close();
@@ -158,22 +158,25 @@ public class ClientFileConnector {
 
 				// Add the file to the user
 				userFile.addFile(location);
-				return true;
+				uploadSuccessful = true;
 			} catch (IOException e) {
 				log.error(e.getMessage());
 				copyErrorToStream("Upload failed. Please contact your server's administrator.", outputStream);
+			} finally {
+				if (!uploadSuccessful) {
+					file.delete();
+				}
 			}
 		} else {
 			log.error("A file generated with FileManager.createFile() was not generated correctly.");
 			copyErrorToStream("Upload failed. Please contact your server's administrator.", outputStream);
 		}
-		file.delete();
-		return false;
+		return uploadSuccessful;
 	}
 
 	/**
 	 * Update a file belonging to the user with the inputStream as its new contents
-	 *
+	 * 
 	 * @param inputStream
 	 *            The contents to fill the file with.
 	 * @param updateRequest
@@ -206,7 +209,7 @@ public class ClientFileConnector {
 
 	/**
 	 * Remove a file belonging to the user
-	 *
+	 * 
 	 * @param removeRequest
 	 *            The request containing the location of the file that needs to be removed
 	 * @param outputStream
@@ -255,7 +258,7 @@ public class ClientFileConnector {
 
 	/**
 	 * Updates the keyfile with the inputStream as its content.
-	 *
+	 * 
 	 * @param inputStream
 	 *            The stream to fill the user's keyfile with
 	 * @param outputStream
